Adding Marci's pre-final draft version of manual
Marci's Notes:

Track changes have been accepted, page numbers have been updated, etc.
I printed the document and did some quality control and again checked
the page number references.  I did not yet convert it to a pdf because
we still don't know about the branding/copyright issue.  I changed the
date on the front of the manual to September 16, 2013 and changed it to
say "Version 1.1" instead of Edition 1 or whatever it used to say.

The other things that I don't want to lose sight of (for now or next
revision) are:
1. The Jobs Tab--does it still exist?  If not, we need to remove
information on page 30, 80, etc.
2. Glossary - there are terms but no definitions for "Differential
Modulation" and "TEA Scores".  I think we should probably delete these,
but I didn't want to do so at this time since clearly at some point
someone felt these needed to be in the glossary.
3. Additional analyses--not all analyses are currently represented for
the Advanced Workflow.